+++ HCMS-Api/Components/DMS/Common/Models/DocumentTypeDto.cs
@@ -9,7 +9,7 @@ public class DocumentType : BaseEntity
     public int Id { get; set; }
 
     // 🔑 Tenant
-    public int CompanyId { get; set; }
+    public Int64 CompanyId { get; set; }
     public Company Company { get; set; } = null!;
 
     public string Code { get; set; } = null!;
@@ -22,7 +22,7 @@ public class DocumentType : BaseEntity
 public class DocumentTypeCreateDto
 {
     // 🔑 Tenant
-    public int CompanyId { get; set; } 
+    public Int64 CompanyId { get; set; } 
     public string Name { get; set; }
     public string? Description { get; set; }
 }
@@ -31,7 +31,7 @@ public class DocumentTypeUpdateDto
 {
 
     // 🔑 Tenant
-    public int CompanyId { get; set; } 
+    public Int64 CompanyId { get; set; } 
     public string Code { get; set; }
     public string Name { get; set; }
 
@@ -43,10 +43,10 @@ public class DocumentTypeUpdateDto
 
 public class DocumentTypeReadDto : AuditableEntity
 {
-    public int Id { get; set; }
+    public Int64 Id { get; set; }
 
     // 🔑 Tenant
-    public int CompanyId { get; set; }
+    public Int64 CompanyId { get; set; }
     public string Company { get; set; } = null!; 
     public string? Description { get; set; }
     public string Code { get; set; }
