+++ HCMS-Api/Components/DMS/Common/Models/TrainingPolicy.cs
@@ -10,7 +10,7 @@ public class TrainingPolicy : AuditableEntity
     public int Id { get; set; }
 
     // 🔑 Tenant
-    public int CompanyId { get; set; }
+    public Int64 CompanyId { get; set; }
     public Company Company { get; set; } = null!;
 
     public int DocumentTypeId { get; set; }
@@ -25,7 +25,7 @@ public class TrainingPolicyReadDto : AuditableEntity
     public int Id { get; set; }
 
     // 🔑 Tenant
-    public int CompanyId { get; set; }
+    public Int64 CompanyId { get; set; }
     public string Company { get; set; } = null!;
 
     public string DocumentTypeCode { get; set; }
@@ -37,7 +37,7 @@ public class TrainingPolicyReadDto : AuditableEntity
 public class TrainingPolicyCreateDto
 {
     // 🔑 Tenant
-    public int CompanyId { get; set; } 
+    public Int64 CompanyId { get; set; } 
     public string DocumentTypeCode { get; set; }
     public bool TrainingRequired { get; set; }
     public int? MinimumScore { get; set; }
@@ -49,7 +49,7 @@ public class TrainingPolicyUpdateDto
     public int Id { get; set; }
 
     // 🔑 Tenant
-    public int CompanyId { get; set; } 
+    public Int64 CompanyId { get; set; } 
 
     public string DocumentTypeCode { get; set; }
     public bool TrainingRequired { get; set; }
