+++ HCMS-Api/Components/DMS/ESS/DocumentAttributeComponent.cs
@@ -121,7 +121,7 @@ public class DocumentAttributeComponent
             return new DocumentAttributeReadDto
             {
                 Id = row.Table.Columns.Contains("Id") ? row.Field<int>("Id") : 0,
-                CompanyId = row.Field<int>("CompanyId"),
+                CompanyId = row.Field<Int64>("CompanyId"),
                 Company = row.Field<string>("Company"),
                 DocumentType = row.Table.Columns.Contains("DocumentType") ? row.Field<string>("DocumentType") : string.Empty,
                 DocumentTypeCode = row.Table.Columns.Contains("DocumentTypeCode") ? row.Field<string>("DocumentTypeCode") : string.Empty,
@@ -242,7 +242,7 @@ public class DocumentAttributeComponent
                 .Select(row => new DocumentAttributeReadDto
                 {
                     Id = row.Table.Columns.Contains("Id") ? row.Field<int>("Id") : 0,
-                    CompanyId = row.Field<int>("CompanyId"),
+                    CompanyId = row.Field<Int64>("CompanyId"),
                     Company = row.Field<string>("Company"),
                     DocumentType = row.Table.Columns.Contains("DocumentType") ? row.Field<string>("DocumentType") : string.Empty,
                     DocumentTypeCode = row.Table.Columns.Contains("DocumentTypeCode") ? row.Field<string>("DocumentTypeCode") : string.Empty,
@@ -335,7 +335,7 @@ public class DocumentAttributeComponent
             return new DocumentAttributeReadDto
             {
                 Id = row.Table.Columns.Contains("Id") ? row.Field<int>("Id") : 0,
-                CompanyId = row.Field<int>("CompanyId"),
+                CompanyId = row.Field<Int64>("CompanyId"),
                 Company = row.Field<string>("Company"),
                 DocumentType = row.Table.Columns.Contains("DocumentType") ? row.Field<string>("DocumentType") : string.Empty,
                 DocumentTypeCode = row.Table.Columns.Contains("DocumentTypeCode") ? row.Field<string>("DocumentTypeCode") : string.Empty,
@@ -385,7 +385,7 @@ public class DocumentAttributeComponent
             return new DocumentAttributeReadDto
             {
                 Id = row.Table.Columns.Contains("Id") ? row.Field<int>("Id") : 0,
-                CompanyId = row.Field<int>("CompanyId"),
+                CompanyId = row.Field<Int64>("CompanyId"),
                 Company = row.Field<string>("Company"),
                 DocumentType = row.Table.Columns.Contains("DocumentType") ? row.Field<string>("DocumentType") : string.Empty,
                 DocumentTypeCode = row.Table.Columns.Contains("DocumentTypeCode") ? row.Field<string>("DocumentTypeCode") : string.Empty,
@@ -468,7 +468,7 @@ public class DocumentAttributeComponent
             return new DocumentAttributeReadDto
             {
                 Id = row.Field<int>("Id"),
-                CompanyId = row.Field<int>("CompanyId"),
+                CompanyId = row.Field<Int64>("CompanyId"),
                 Company = row.Field<string>("Company"),
                 DocumentTypeCode = row.Field<string>("DocumentTypeCode"),
                 ControlLabel = row.Field<string>("ControlLabel"),
